refactor: change command folder name
